Move edit code from libimagrt
