+++ libimagentryedit/src/edit.rs
