+++ libimagentryedit/src/lib.rs
@@ -1,6 +1,1 @@
-#[cfg(test)]
-mod tests {
-    #[test]
-    fn it_works() {
-    }
-}
+pub mod edit;
