+++ libimagrt/src/lib.rs
@@ -32,7 +32,6 @@ extern crate libimagutil;
 mod configuration;
 mod logger;
 
-pub mod edit;
 pub mod error;
 pub mod runtime;
 pub mod setup;
